backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
pymongo
psycopg2-binary
requests
psycopg2
passlib[bcrypt]
python-jose
gunicorn